+++ apps/avatax/src/modules/taxes/tax-error.ts
@@ -5,6 +5,9 @@ export const TaxIncompleteWebhookPayloadError = ExpectedError.subclass(
   "TaxIncompleteWebhookPayloadError",
   {
     props: {
+      /**
+       * TODO: Remove this field
+       */
       sentrySeverity: "warning",
     } as CommonErrorProps,
   },